+++ hTech/BoundingPolygon.h
@@ -11,11 +11,11 @@ private:
 	std::vector<Vector2> mTransformedPoints;
 
 public:
-	int PointCount = 0;
+	size_t PointCount = 0;
 
 	BoundingPolygon(Transform& transform, float& rotation, const std::vector<Vector2> points, const int& pointCount);
 
-	virtual void Update(double DeltaTime);
+	virtual void Update(float DeltaTime);
 	virtual void Render(SDL_Renderer& renderer);
 
 	virtual Vector2 FindFurthestPoint(Vector2 direction) const;
